feat:それっぽく再生成実装
