+++ botmain.js
@@ -19,6 +19,7 @@ global.client = new Client({
 const config = require('./config.json');
 const system = require('./functions/logsystem.js');
 const help = require("./functions/help.js");
+const makeTxt = require("./functions/makeTxt.js");
 
 //スラッシュコマンド登録
 const commandsPath = path.join(__dirname, 'commands');
@@ -93,71 +94,32 @@ client.on(Events.InteractionCreate, async interaction => {
 
 
 client.on('messageCreate', async message => {
-    const nl = new RegExp("\n");
-    const twitter = new RegExp("https?://twitter.com/");
-    const xcom = new RegExp("https?://x.com/");
-    const vxtwitter = new RegExp("https?://vxtwitter.com/");
-    const embedLink = new RegExp("[-_.!~*'a-zA-Z0-9;?:&=+$,%#]");
-    const specialChar = new RegExp("[^-_.!~*'a-zA-Z0-9;?:&=+$,%#/]");
-
-
-    if(twitter.test(message.content) || xcom.test(message.content)){
-        let content = message.content;
-        while(specialChar.test(content)){
-            content = content.replace(specialChar,"");
+    const msg = makeTxt.make(message.content);
+    if(msg!==""){
+        try{
+            const embed = await message.channel.send(msg);
+            await embed.react('❌');
         }
-        while(nl.test(content)){
-            content = content.replace(nl,"");
-        }
-        while(vxtwitter.test(content)){
-            content = content.replace(twitter,"xxxxxxxxxxxxvxtwitterx");
-        }
-        while(twitter.test(content) || xcom.test(content)){
-            content = content.replace(twitter,"https://vxtwitter.com/");
-            content = content.replace(xcom,"https://vxtwitter.com/");
-        }
-
-        const lines = content.split(/\r\n|\r|\n/);
-        let url = [];
-        for(let l of lines){
-            url = url.concat(l.split(vxtwitter));
-        }
-
-        let msg = "";
-        let num=1;
-        for(let i=1;i<url.length;i++){
-            if(embedLink.test(url[i])){
-                msg += `[URL${num}](https://vxtwitter.com/${url[i]})\n`;
-                num++;
+        catch(err){
+            if(err.code!==50013){
+                try{
+                    let guild,channel;
+                    if(!message.guildId) {
+                        guild = {name:"ダイレクトメッセージ",id:"---"};
+                        channel = {name:"---",id:"---"};
+                    }
+                    else{
+                        guild = client.guilds.cache.get(message.guildId) ?? await client.guilds.fetch(message.guildId);
+                        channel = client.channels.cache.get(message.channelId) ?? await client.channels.fetch(message.channelId);
+                    }
+                    await system.error(`権限不足等の要因でメッセージを送信できませんでした\`\`\`\nギルド　　：${guild.name}\n(ID:${guild.id})\n\nチャンネル：${channel.name}\n(ID:${channel.id})\n\nID　　　　：${message.id}\n\nユーザ　　：${message.author.globalName}${(message.author.discriminator==="0")?"":`#${message.author.discriminator}`}\n(ID:${message.author.id})\`\`\``,err,"メッセージ送信失敗");
+                }
+                catch{
+                    await system.error(`権限不足等の要因でメッセージを送信できませんでした`,err,"メッセージ送信失敗");
+                }
             }
         }
 
-        if(msg!==""){
-            try{
-                const embed = await message.channel.send(msg);
-                await embed.react('❌');
-            }
-            catch(err){
-               if(err.code!==50013){
-                   try{
-                       let guild,channel;
-                       if(!message.guildId) {
-                           guild = {name:"ダイレクトメッセージ",id:"---"};
-                           channel = {name:"---",id:"---"};
-                       }
-                       else{
-                           guild = client.guilds.cache.get(message.guildId) ?? await client.guilds.fetch(message.guildId);
-                           channel = client.channels.cache.get(message.channelId) ?? await client.channels.fetch(message.channelId);
-                       }
-                       await system.error(`権限不足等の要因でメッセージを送信できませんでした\`\`\`\nギルド　　：${guild.name}\n(ID:${guild.id})\n\nチャンネル：${channel.name}\n(ID:${channel.id})\n\nID　　　　：${message.id}\n\nユーザ　　：${message.author.globalName}${(message.author.discriminator==="0")?"":`#${message.author.discriminator}`}\n(ID:${message.author.id})\`\`\``,err,"メッセージ送信失敗");
-                   }
-                   catch{
-                       await system.error(`権限不足等の要因でメッセージを送信できませんでした`,err,"メッセージ送信失敗");
-                   }
-               }
-            }
-
-        }
     }
 });
 
@@ -167,6 +129,33 @@ client.on('messageReactionAdd', async reaction => {
             await reaction.message.delete();
         } catch (err) {
             await system.error(`メッセージの削除に失敗しました`, err, "メッセージ削除失敗");
+        }
+    }
+    else if(reaction._emoji.name === '⭕'){
+        const msg = makeTxt.make(reaction.message.content);
+        if(msg!==""){
+            try{
+                await reaction.message.reply(msg);
+            }
+            catch(err){
+                if(err.code!==50013){
+                    try{
+                        let guild,channel;
+                        if(!message.guildId) {
+                            guild = {name:"ダイレクトメッセージ",id:"---"};
+                            channel = {name:"---",id:"---"};
+                        }
+                        else{
+                            guild = client.guilds.cache.get(message.guildId) ?? await client.guilds.fetch(message.guildId);
+                            channel = client.channels.cache.get(message.channelId) ?? await client.channels.fetch(message.channelId);
+                        }
+                        await system.error(`権限不足等の要因でメッセージを送信できませんでした\`\`\`\nギルド　　：${guild.name}\n(ID:${guild.id})\n\nチャンネル：${channel.name}\n(ID:${channel.id})\n\nID　　　　：${message.id}\n\nユーザ　　：${message.author.globalName}${(message.author.discriminator==="0")?"":`#${message.author.discriminator}`}\n(ID:${message.author.id})\`\`\``,err,"メッセージ送信失敗");
+                    }
+                    catch{
+                        await system.error(`権限不足等の要因でメッセージを送信できませんでした`,err,"メッセージ送信失敗");
+                    }
+                }
+            }
         }
     }
 })
